requirements.txt
djangorestframework
requests
django-taggit
django-taggit-serializer
djangorestframework-jwt
django
pyyaml
coverage